+++ MatrixSDK/Crypto/Algorithms/Megolm/MXMegolmDecryption.m
@@ -240,14 +240,22 @@
     [self retryDecryption:senderKey sessionId:content[@"session_id"]];
 }
 
-- (void)importRoomKey:(MXMegolmSessionData *)session
+- (BOOL)importRoomKey:(MXMegolmSessionData *)session backUp:(BOOL)backUp
 {
-    [olmDevice importInboundGroupSession:session];
+    BOOL imported = [olmDevice importInboundGroupSession:session];
+    if (imported)
+    {
+        // Do not back up the key if it comes from a backup recovery
+        if (backUp)
+        {
+            [crypto.backup maybeSendKeyBackup];
+        }
 
-    [crypto.backup maybeSendKeyBackup];
+        // Have another go at decrypting events sent with this session
+        [self retryDecryption:session.senderKey sessionId:session.sessionId];
+    }
 
-    // Have another go at decrypting events sent with this session
-    [self retryDecryption:session.senderKey sessionId:session.sessionId];
+    return imported;
 }
 
 - (BOOL)hasKeysForKeyRequest:(MXIncomingRoomKeyRequest*)keyRequest